fix: PNG デコーダーの bytes_per_pixel を u32 に修正
filter.applyFilters() が u32 を期待しているため、型を合わせる。
すべてのテストが正常に実行される。

+++ libs/png-decoder/src/lib.zig
@@ -68,7 +68,7 @@ pub fn decodePNG(allocator: std.mem.Allocator, file_data: []const u8) DecodingEr
 
     // Determine bytes per pixel based on color type
     const color_type: ColorType = @enumFromInt(ihdr.color_type);
-    const bytes_per_pixel: usize = switch (color_type) {
+    const bytes_per_pixel: u32 = switch (color_type) {
         .grayscale => 1,
         .rgb => 3,
         .grayscale_alpha => 2,
